feat(Fields): fix listener and filter Document
